Popup and Editor
Added Component editor.

Popup menu now closes on unfocus.

+++ ElecDev_Graphics_Application/Source/Graphics/OpenGL/Design2DEngineGL/Design2D_Engine.cpp
@@ -94,7 +94,7 @@ void Design2DEngineGL::renderLoop()
 //  Events.																																	   //
 //=============================================================================================================================================//
 
-void Design2DEngineGL::resizeEvent(int width, int height)
+void Design2DEngineGL::resizeEvent(float width, float height)
 {
 	// Call base engine resize event.
 	Base2DEngineGL::resizeEvent(width, height);
